生産計画APIの変更に対処
