+++ backend/apps/prod_flow/urls.py
@@ -1,10 +1,10 @@
 from django.urls import path
-from apps.prod_flow.views import ProductionPlanView, ProductionPlanWithRecordView, ProductionPlanWithDetailsUpdateView, ProductionPlanDeleteView
+from apps.prod_flow.views import ProductionPlanView, ProductionPlanWithRecordView, ProductionPlanWithRecordDetailView
 
 
 urlpatterns = [
     path("plans/", ProductionPlanView.as_view(), name="production-plan"),
     path("plan_with_records/", ProductionPlanWithRecordView.as_view(), name="production-plan-with-record"),
-    path("plans/<int:pk>/update/", ProductionPlanWithDetailsUpdateView.as_view(), name="production-plan-update"),
-    path('plans/<int:pk>/delete/', ProductionPlanDeleteView.as_view(), name='production-plan-delete'),
+    path("plan_with_records/<int:pk>/", ProductionPlanWithRecordDetailView.as_view(), name="production-plan-with-record-detail"),
+    # path('plan_with_records/<int:pk>/delete/', ProductionPlanDeleteView.as_view(), name='production-plan-delete'),
 ]
