+++ backend/apps/prod_flow/views/__init__.py
@@ -1,4 +1,3 @@
 from .production_plan_with_record_view import ProductionPlanWithRecordView
 from .production_plan_view import ProductionPlanView
-from .production_plan_with_details_update_view import ProductionPlanWithDetailsUpdateView
-from .production_plan_delete_view import ProductionPlanDeleteView
+from .production_plan_with_record_detail_view import ProductionPlanWithRecordDetailView
