+++ backend/apps/prod_flow/views/production_plan_with_record_detail_view.py
@@ -7,9 +7,11 @@ from apps.staff_hub.permission import HasUserPermissionObject
 from apps.prod_flow.common import check_prod_flow_edit_permission
 from apps.prod_flow.models.production_plan import ProductionPlan
 from django.shortcuts import get_object_or_404
+from django.utils import timezone
+from apps.prod_flow.models.production_plan_detail import ProductionPlanDetail
 
 
-class ProductionPlanWithDetailsUpdateView(APIView):
+class ProductionPlanWithRecordDetailView(APIView):
     permission_classes = [permissions.IsAuthenticated, HasUserPermissionObject]
 
     @extend_schema(
@@ -28,3 +30,22 @@ class ProductionPlanWithDetailsUpdateView(APIView):
         plan = serializer.save()
 
         return Response(ProductionPlanWithDetailsSerializer(plan).data, status=status.HTTP_200_OK)
+
+    @extend_schema(
+        tags=["production"],
+        description="生産計画と詳細を論理削除する",
+        responses={204: None}
+    )
+    def delete(self, request, pk):
+        check_prod_flow_edit_permission(request)
+
+        plan = get_object_or_404(ProductionPlan, pk=pk)
+
+        plan.deleted_at = timezone.now()
+        plan.save()
+
+        ProductionPlanDetail.objects.filter(production_plan=plan).update(
+            deleted_at=timezone.now()
+        )
+
+        return Response(status=status.HTTP_204_NO_CONTENT)
